rename class to FullCalendarComponent. move files around
