+++ projects/fullcalendar/src/lib/fullcalendar-options.ts
