+++ projects/fullcalendar/src/lib/fullcalendar.component.spec.ts
@@ -1,20 +1,20 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { CalendarComponent } from './calendar.component';
+import { FullCalendarComponent } from './fullcalendar.component';
 
-describe('CalendarComponent', () => {
-  let component: CalendarComponent;
-  let fixture: ComponentFixture<CalendarComponent>;
+describe('FullCalendarComponent', () => {
+  let component: FullCalendarComponent;
+  let fixture: ComponentFixture<FullCalendarComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ CalendarComponent ]
+      declarations: [ FullCalendarComponent ]
     })
     .compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(CalendarComponent);
+    fixture = TestBed.createComponent(FullCalendarComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
