+++ projects/fullcalendar/src/public_api.ts
@@ -2,5 +2,5 @@
  * Public API Surface of fullcalendar
  */
 
-export * from './lib/calendar/calendar.component';
+export * from './lib/fullcalendar.component';
 export * from './lib/fullcalendar.module';
